+++ src/date/fallback-format/fallbackFormat.spec.js
@@ -1,0 +1,61 @@
+import { getFallbackFormat } from '.';
+
+describe('Date format fallback', () => {
+  const date = new Date(Date.UTC(2018, 11, 1));
+
+  it('formats DMY by default', () => {
+    const opts = { timeZone: 'UTC' };
+    expect(getFallbackFormat(date, opts)).toBe('1/12/2018');
+  });
+
+  it('formats W', () => {
+    const opts = { timeZone: 'UTC', weekday: true };
+    expect(getFallbackFormat(date, opts)).toBe('Sat');
+  });
+
+  it('formats W, D', () => {
+    const opts = { timeZone: 'UTC', weekday: true, day: true };
+    expect(getFallbackFormat(date, opts)).toBe('Sat, 1');
+  });
+
+  it('formats W, DM', () => {
+    const opts = { timeZone: 'UTC', weekday: true, month: true, day: true };
+    expect(getFallbackFormat(date, opts)).toBe('Sat, 1/12');
+  });
+
+  it('formats W, DMY', () => {
+    const opts = { timeZone: 'UTC', weekday: true, year: true, month: true, day: true };
+    expect(getFallbackFormat(date, opts)).toBe('Sat, 1/12/2018');
+  });
+
+  it('formats Y', () => {
+    const opts = { timeZone: 'UTC', year: true };
+    expect(getFallbackFormat(date, opts)).toBe('2018');
+  });
+
+  it('formats MY', () => {
+    const opts = { timeZone: 'UTC', year: true, month: true };
+    expect(getFallbackFormat(date, opts)).toBe('12/2018');
+  });
+
+  it('formats DMY', () => {
+    const opts = { timeZone: 'UTC', year: true, month: true, day: true };
+    expect(getFallbackFormat(date, opts)).toBe('1/12/2018');
+  });
+
+  it('formats DM', () => {
+    const opts = { timeZone: 'UTC', month: true, day: true };
+    expect(getFallbackFormat(date, opts)).toBe('1/12');
+  });
+
+  it('formats D', () => {
+    const opts = { timeZone: 'UTC', day: true };
+    expect(getFallbackFormat(date, opts)).toBe('1');
+  });
+
+  it('works for non-UTC dates', () => {
+    const date = new Date('2018-12-01');
+    const expectedOutput = `${date.getDate()}/${date.getMonth() + 1}/${date.getFullYear()}`;
+    expect(getFallbackFormat(date)).toBe(expectedOutput);
+  });
+});