+++ src/date/support-detection/index.js
@@ -1,0 +1,2 @@
+export { isToLocaleDateStringSupported } from './toLocaleDateString';
+export { isSelectedLocaleSupported } from './selectedLocale';