+++ src/index.js
@@ -1,3 +1,4 @@
 export * from './currency';
 export * from './rate';
 export * from './percentage';
+export * from './date';